[FEAT] 테스트 코드 작성 (#73)

+++ src/main/java/org/chzzk/howmeet/domain/regular/room/service/RoomService.java
@@ -73,6 +73,9 @@ public class RoomService {
 
     @Transactional
     public void deleteRoom(Long roomId) {
+        if (!roomRepository.existsById(roomId)) {
+            throw new RuntimeException("Invalid room ID");
+        }
         roomRepository.deleteById(roomId);
     }
